+++ course/format/grid/Changes.md
@@ -1,5 +1,10 @@
 History
 =============
+Version 3.10.1.4 - 01/07/21
+============================
+ 1. Fix 'Exception when using course import function' - #144.
+ 2. Fix 'Course restore fails with cannotgetimagesforcourse error' - #145.
+
 Version 3.10.1.3 - 02/06/21
 ============================
 Due to the changes, this is a 'release candidate' release, therefore it is imperative that you test on a test server