+++ course/format/topcoll/Changes.md
@@ -1,10 +1,16 @@
 ﻿Version Information
 ===================
-Version 3.10.1.2 - TBD
-----------------------
+Version 3.10.1.2 - 02/07/2021
+-----------------------------
   1. Fix 'PHP Notice:  Trying to access array offset on value of type null' when accessing a single section page on PHP 7.4.
   2. Reset language improvements in line with global default tracking as implemented in 3.10.0.2.
   3. Improved backup restore with numsections.
+  4. Added 'Toggle blocks location' functionality, thanks to John Joubert (https://github.com/John-Joubert) for the patch - #100.
+  5. Fix 'Section content is in a column when using horizontal columns on a single section page'.
+  6. Fix 'Column orientation combo box strings'.
+  7. Improve activity meta from information gained on #101.
+  8. No longer need 'config.php' because of #100.
+  9. New 'Accessible back to main course link' - #66.
 
 Version 3.10.1.1 - 07/03/2021
 -----------------------------